Working Picture code
When Pin 27 goes High, a picture is taken with the camera. The basis
behind this is that the scanner clamp will be a switch and allow pin 27
to go high and take a picture (with a possible delay) when the page is
"clamped on". Will only take a picture once per high pulse, and waits
for it to go low, then high again (Schmitt trigger).